+++ src/main/java/envision/engine/windows/windowObjects/advancedObjects/menuBar/WindowMenuBar.java
@@ -34,7 +34,7 @@ public class WindowMenuBar extends WindowObject {
     }
     
     @Override
-    public void drawObject(long dt, int mXIn, int mYIn) {
+    public void drawObject(float dt, int mXIn, int mYIn) {
         drawRect(EColors.black);
         drawRect(EColors.pdgray, 1);
